🧩 Refactored ChatInputArea to integrate image upload and feature selection.
🆕 Developed MessageItem component to render messages with support for code blocks and markdown.

🆕 Introduced useChatScroll hook for better scroll management in chat.
